djangoApp/damnificados/requirements.txt
django
mysqlclient
gunicorn
djangorestframework==3.6.4
requests